Change params format to use reporter_params
App was crashing when clicking on a phone# or name in 'views/reports/index'. The corresponding controller, 'reporters_controller#show,' takes in the private method reporter_params which requires a 'reporter' resource and permits specific values. To fix the bug, it required changing the format of params in 'views/reports/index.'

Also, address was added as a valid input for reporter_params, allowing filtering by address on 'views/reporters/show'